+++ Cerberus/Core/Utility/InputManager/InputManager.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Vector3.h"
+#include "Utility/Vector3.h"
 
 namespace Inputs
 {
